feat:add some details and improvements for the team offers feature
