+++ src/components/commun/OperationsHoverButton.jsx
@@ -3,10 +3,10 @@ import {
   HoverCardContent,
   HoverCardTrigger,
 } from "@/components/ui/hover-card";
-import { FilePen, Pencil } from "lucide-react";
+
 import { Button } from "../ui/button";
 import { Link } from "react-router";
-function EditHoverButton({ children, path }) {
+function OperationsHoverButton({ children, path, icon }) {
   return (
     <HoverCard>
       <HoverCardTrigger asChild>
@@ -14,16 +14,14 @@ function EditHoverButton({ children, path }) {
           className="fixed bottom-3 right-3 rounded-full w-12 h-12 lg:w-14 lg:h-14 ring-offset-2 ring-4 hover:ring-primary transition-all duration-300 cursor-pointer"
           asChild
         >
-          <Link to={path}>
-            <Pencil />
-          </Link>
+          <Link to={path}>{icon}</Link>
         </Button>
       </HoverCardTrigger>
       <HoverCardContent className="flex items-center justify-center gap-2 text-primary w-auto">
-        <FilePen /> {children}
+        {icon} {children}
       </HoverCardContent>
     </HoverCard>
   );
 }
 
-export default EditHoverButton;
+export default OperationsHoverButton;
