+++ src/modules/student/features/useAddTeamOffer.js
@@ -1,18 +1,23 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import toast from "react-hot-toast";
 import addTeamOffer from "../api/apiAddTeamOffer";
+import { useNavigate } from "react-router";
 
-export const useaddTeamOffer = () => {
+export function useAddTeamOffer() {
+  const navigate = useNavigate();
   const queryClient = useQueryClient();
   const { mutate: addTeamOfferMutation, isPending: isAdding } = useMutation({
     mutationFn: (data) => addTeamOffer(data),
     onSuccess: () => {
+      queryClient.invalidateQueries(["teams"]);
       toast.success("Offer submitted successfully!");
-      queryClient.invalidateQueries(["teams"]); 
+      navigate("/student/team-offers");
     },
     onError: (error) => {
       console.error("Error submitting offer:", error);
-      toast.error(error.message || "Failed to submit offer. Please try again later.");
+      toast.error(
+        error.message || "Failed to submit offer. Please try again later."
+      );
     },
   });
 
@@ -20,4 +25,4 @@ export const useaddTeamOffer = () => {
     addTeamOffer: addTeamOfferMutation,
     isAdding,
   };
-};
+}
